Use plain retirement contribution outputs
