+++ policyengine_us/parameters/gov/contrib/crfb/surtax/increased_base/sources.yaml
@@ -2,9 +2,9 @@ description: Additional income sources for CRFB surtax expanded base beyond IRS 
 values:
   2010-01-01:
     # Retirement contributions (IRA, 401k, 403b)
-    - capped_traditional_ira_contributions
-    - capped_traditional_401k_contributions
-    - capped_traditional_403b_contributions
+    - traditional_ira_contributions
+    - traditional_401k_contributions
+    - traditional_403b_contributions
     # HSA contributions (no Archer MSA variable exists)
     - health_savings_account_ald
     # Deductible student loan interest
