+++ policyengine_us/parameters/gov/irs/ald/deductions.yaml
@@ -12,7 +12,7 @@ values:
     - health_savings_account_ald
     - self_employed_health_insurance_ald
     - self_employed_pension_contribution_ald
-    - capped_traditional_ira_contributions
+    - traditional_ira_contributions
     - qualified_adoption_assistance_expense
     - us_bonds_for_higher_ed
     - specified_possession_income
@@ -28,7 +28,7 @@ values:
     - health_savings_account_ald
     - self_employed_health_insurance_ald
     - self_employed_pension_contribution_ald
-    - capped_traditional_ira_contributions
+    - traditional_ira_contributions
     - qualified_adoption_assistance_expense
     - us_bonds_for_higher_ed
     - specified_possession_income
@@ -43,7 +43,7 @@ values:
     - health_savings_account_ald
     - self_employed_health_insurance_ald
     - self_employed_pension_contribution_ald
-    - capped_traditional_ira_contributions
+    - traditional_ira_contributions
     - qualified_adoption_assistance_expense
     - us_bonds_for_higher_ed
     - specified_possession_income
@@ -58,7 +58,7 @@ values:
     - health_savings_account_ald
     - self_employed_health_insurance_ald
     - self_employed_pension_contribution_ald
-    - capped_traditional_ira_contributions
+    - traditional_ira_contributions
     - qualified_adoption_assistance_expense
     - us_bonds_for_higher_ed
     - specified_possession_income
