+++ policyengine_us/parameters/gov/irs/credits/retirement_saving/qualified_retirement_savings_contributions.yaml
@@ -1,23 +1,23 @@
 description: The IRA sums the following elements as qualified retirement savings contributions when computing the saver's credit.  
 values:
   2018-01-01:
-    - capped_traditional_ira_contributions
-    - capped_roth_ira_contributions
-    - capped_traditional_401k_contributions
-    - capped_traditional_403b_contributions
-    - capped_roth_401k_contributions
-    - capped_roth_403b_contributions
-    - capped_self_employed_pension_contributions
+    - traditional_ira_contributions
+    - roth_ira_contributions
+    - traditional_401k_contributions
+    - traditional_403b_contributions
+    - roth_401k_contributions
+    - roth_403b_contributions
+    - self_employed_pension_contributions
     - able_contributions_person
 
   2026-01-01:
-    - capped_traditional_ira_contributions
-    - capped_roth_ira_contributions
-    - capped_traditional_401k_contributions
-    - capped_traditional_403b_contributions
-    - capped_roth_401k_contributions
-    - capped_roth_403b_contributions
-    - capped_self_employed_pension_contributions
+    - traditional_ira_contributions
+    - roth_ira_contributions
+    - traditional_401k_contributions
+    - traditional_403b_contributions
+    - roth_401k_contributions
+    - roth_403b_contributions
+    - self_employed_pension_contributions
 
 metadata:
   unit: list
