+++ policyengine_us/parameters/gov/local/ca/riv/general_relief/income/deductions/sources.yaml
@@ -5,7 +5,7 @@ values:
     - employee_social_security_tax
     - employee_medicare_tax
     - additional_medicare_tax
-    - capped_traditional_ira_contributions
+    - traditional_ira_contributions
 
 metadata:
   unit: list
