+++ policyengine_us/parameters/gov/states/al/tax/income/agi/deductions.yaml
@@ -2,13 +2,13 @@ description: Alabama provides these deductions for the state adjusted gross inco
 values:
   2021-01-01:
   # IRA distributions prior to 1982 are not deductible.
-    - capped_traditional_ira_contributions # Line 1
+    - traditional_ira_contributions # Line 1
     - alimony_expense # Line 4
     - self_employed_health_insurance_ald # Line 7
     - us_govt_interest
   2023-01-01:
   # IRA distributions prior to 1982 are not deductible.
-    - capped_traditional_ira_contributions # Line 1
+    - traditional_ira_contributions # Line 1
     - alimony_expense # Line 4
     - self_employed_health_insurance_ald # Line 7
     - us_govt_interest
