+++ policyengine_us/parameters/gov/states/ia/tax/income/income_adjustments/sources.yaml
@@ -1,7 +1,7 @@
 description: Iowa subtracts these sources from gross income.
 values:
   2021-01-01:
-    - capped_traditional_ira_contributions
+    - traditional_ira_contributions
     - self_employment_tax_ald_person
     - alimony_expense
     - ia_pension_exclusion
