+++ policyengine_us/parameters/gov/states/ut/tax/property/homeowner_renter_relief/nontaxable_income/sources.yaml
@@ -8,10 +8,10 @@ values:
     - tax_exempt_interest_income
     - tax_exempt_pension_income
     - tax_exempt_retirement_distributions
-    - capped_traditional_401k_contributions
-    - capped_traditional_403b_contributions
-    - capped_traditional_ira_contributions
-    - capped_self_employed_pension_contributions
+    - traditional_401k_contributions
+    - traditional_403b_contributions
+    - traditional_ira_contributions
+    - self_employed_pension_contributions
     - veterans_benefits
     - workers_compensation
 metadata:
