+++ policyengine_us/tests/policy/baseline/contrib/biden/budget_2025/medicare_and_investment_tax_increase.yaml
@@ -28,7 +28,7 @@
     people:
       head:
         employment_income: 450_000
-        traditional_401k_contributions: 50_000
+        traditional_401k_contributions_desired: 50_000
     tax_units:
       tax_unit:
         members: [head]
