+++ policyengine_us/tests/policy/baseline/gov/irs/income/taxable_income/adjusted_gross_income/above_the_line_deductions/self_employment_income.yaml
@@ -18,7 +18,7 @@
   period: 2024
   input:
     sstb_self_employment_income: 8_000
-    self_employed_pension_contributions: 9_000
+    self_employed_pension_contributions_desired: 9_000
   output:
     self_employed_pension_contribution_ald_person: 8_000
