+++ policyengine_us/tests/policy/baseline/gov/irs/income/taxable_income/adjusted_gross_income/irs_gross_income/roth_contributions_not_pre_tax.yaml
@@ -2,7 +2,7 @@
   period: 2025
   input:
     employment_income: 100_000
-    roth_401k_contributions: 20_000
+    roth_401k_contributions_desired: 20_000
   output:
     pre_tax_contributions: 0
     irs_employment_income: 100_000
@@ -11,7 +11,7 @@
   period: 2025
   input:
     employment_income: 100_000
-    roth_403b_contributions: 15_000
+    roth_403b_contributions_desired: 15_000
   output:
     pre_tax_contributions: 0
     irs_employment_income: 100_000
@@ -21,12 +21,12 @@
   absolute_error_margin: 0.01
   input:
     employment_income: 100_000
-    traditional_401k_contributions: 10_000
-    roth_401k_contributions: 10_000
-    traditional_403b_contributions: 5_000
-    roth_403b_contributions: 5_000
+    traditional_401k_contributions_desired: 10_000
+    roth_401k_contributions_desired: 10_000
+    traditional_403b_contributions_desired: 5_000
+    roth_403b_contributions_desired: 5_000
   output:
-    capped_traditional_401k_contributions: 7_833.33
-    capped_traditional_403b_contributions: 3_916.67
+    traditional_401k_contributions: 7_833.33
+    traditional_403b_contributions: 3_916.67
     pre_tax_contributions: 11_750
     irs_employment_income: 88_250
