+++ policyengine_us/tests/policy/baseline/gov/irs/tax/federal_income/additional_medicare_tax.yaml
@@ -22,7 +22,7 @@
     people:
       person1:
         employment_income: 250_000
-        traditional_401k_contributions: 50_000
+        traditional_401k_contributions_desired: 50_000
     tax_units:
       tax_unit:
         members: [person1]
