+++ policyengine_us/tests/policy/baseline/gov/states/ma/tax/income/adjusted_gross_income/ma_part_b_agi.yaml
@@ -4,6 +4,6 @@
   input:
     state_code: MA
     employment_income: 10_000
-    traditional_ira_contributions: 1_000
+    traditional_ira_contributions_desired: 1_000
   output:
     ma_part_b_agi: 10_000
