+++ policyengine_us/tests/policy/baseline/gov/territories/pr/tax/income/taxable_income/deductions/retirement/pr_retirement_deduction.yaml
@@ -2,8 +2,8 @@
   period: 2024
   input:
     pr_agi_person: 10_000
-    traditional_ira_contributions: 3_001
-    roth_ira_contributions: 2_000
+    traditional_ira_contributions_desired: 3_001
+    roth_ira_contributions_desired: 2_000
     state_code: PR
   output:
     pr_retirement_deduction: 5_000
@@ -12,8 +12,8 @@
   period: 2024
   input:
     pr_agi_person: 1_300
-    traditional_ira_contributions: 500
-    roth_ira_contributions: 300
+    traditional_ira_contributions_desired: 500
+    roth_ira_contributions_desired: 300
     state_code: PR
   output:
     pr_retirement_deduction: 800
@@ -22,8 +22,8 @@
   period: 2024
   input:
     pr_agi_person: 4_000
-    traditional_ira_contributions: 500
-    roth_ira_contributions: 301
+    traditional_ira_contributions_desired: 500
+    roth_ira_contributions_desired: 301
     state_code: PR
   output:
     pr_retirement_deduction: 801
@@ -32,8 +32,8 @@
   period: 2024
   input:
     pr_agi_person: 4_000
-    traditional_ira_contributions: 5000
-    roth_ira_contributions: 3000
+    traditional_ira_contributions_desired: 5000
+    roth_ira_contributions_desired: 3000
     state_code: PR
   output:
     pr_retirement_deduction: 4_000
