+++ policyengine_us/tests/policy/baseline/household/expense/retirement/elective_deferral_limit.yaml
@@ -1,29 +1,25 @@
-- name: Case 1, legacy desired deferrals below the elective deferral limit.
+- name: Case 1, desired deferrals below the elective deferral limit.
   period: 2025
   input:
     age: 35
-    traditional_401k_contributions: 12_000
-    roth_401k_contributions: 8_000
+    traditional_401k_contributions_desired: 12_000
+    roth_401k_contributions_desired: 8_000
   output:
     elective_deferral_limit: 23_500
     traditional_401k_contributions: 12_000
     roth_401k_contributions: 8_000
-    capped_traditional_401k_contributions: 12_000
-    capped_roth_401k_contributions: 8_000
 
-- name: Case 2, legacy deferral inputs above the base elective deferral limit.
+- name: Case 2, desired deferral inputs above the base elective deferral limit.
   period: 2025
   absolute_error_margin: 0.01
   input:
     age: 35
-    traditional_401k_contributions: 20_000
-    roth_401k_contributions: 10_000
+    traditional_401k_contributions_desired: 20_000
+    roth_401k_contributions_desired: 10_000
   output:
     elective_deferral_limit: 23_500
-    traditional_401k_contributions: 20_000
-    roth_401k_contributions: 10_000
-    capped_traditional_401k_contributions: 15_666.67
-    capped_roth_401k_contributions: 7_833.33
+    traditional_401k_contributions: 15_666.67
+    roth_401k_contributions: 7_833.33
     pre_tax_contributions: 15_666.67
 
 - name: Case 3, desired inputs use the enhanced catch-up limit across 401(k) and 403(b) deferrals.
@@ -37,10 +33,10 @@
     roth_403b_contributions_desired: 5_000
   output:
     elective_deferral_limit: 34_750
-    capped_traditional_401k_contributions: 17_375
-    capped_roth_401k_contributions: 8_687.50
-    capped_traditional_403b_contributions: 4_343.75
-    capped_roth_403b_contributions: 4_343.75
+    traditional_401k_contributions: 17_375
+    roth_401k_contributions: 8_687.50
+    traditional_403b_contributions: 4_343.75
+    roth_403b_contributions: 4_343.75
 
 - name: Case 4, SECURE 2.0 enhanced catch-up applies from ages 60 through 63.
   period: 2025
@@ -61,63 +57,62 @@
   output:
     elective_deferral_limit: [23_500, 31_000, 31_000, 34_750, 34_750, 31_000]
 
-- name: Case 5, capped traditional deferrals reduce income tax wages.
+- name: Case 5, traditional deferrals reduce income tax wages.
   period: 2025
   input:
     age: 40
     employment_income: 100_000
-    traditional_401k_contributions: 50_000
+    traditional_401k_contributions_desired: 50_000
   output:
-    traditional_401k_contributions: 50_000
-    capped_traditional_401k_contributions: 23_500
+    traditional_401k_contributions: 23_500
     pre_tax_contributions: 23_500
     irs_employment_income: 76_500
 
-- name: Case 6, capped traditional deferrals do not reduce FICA wages.
+- name: Case 6, traditional deferrals do not reduce FICA wages.
   period: 2025
   input:
     age: 40
     employment_income: 100_000
-    traditional_401k_contributions: 50_000
+    traditional_401k_contributions_desired: 50_000
   output:
-    capped_traditional_401k_contributions: 23_500
+    traditional_401k_contributions: 23_500
     payroll_tax_gross_wages: 100_000
 
-- name: Case 7, capped Roth deferrals count for the saver's credit.
+- name: Case 7, Roth deferrals count for the saver's credit.
   period: 2025
   input:
     age: 40
-    traditional_401k_contributions: 30_000
-    roth_401k_contributions: 17_000
+    traditional_401k_contributions_desired: 30_000
+    roth_401k_contributions_desired: 17_000
     retirement_distributions: 0
   output:
-    capped_traditional_401k_contributions: 15_000
-    capped_roth_401k_contributions: 8_500
+    traditional_401k_contributions: 15_000
+    roth_401k_contributions: 8_500
     savers_credit_qualified_contributions: 23_500
 
 - name: Case 8, the 2026 standard catch-up applies from ages 50 through 59.
   period: 2026
   input:
     age: 55
-    traditional_401k_contributions: 32_500
+    traditional_401k_contributions_desired: 32_500
   output:
     elective_deferral_limit: 32_500
-    capped_traditional_401k_contributions: 32_500
+    traditional_401k_contributions: 32_500
 
 - name: Case 9, the SECURE 2.0 enhanced catch-up applies in 2026.
   period: 2026
   input:
     age: 62
-    traditional_401k_contributions: 35_750
+    traditional_401k_contributions_desired: 35_750
   output:
     elective_deferral_limit: 35_750
-    capped_traditional_401k_contributions: 35_750
+    traditional_401k_contributions: 35_750
 
 - name: Case 10, age 64 returns to the standard catch-up in 2026.
   period: 2026
   input:
     age: 64
-    traditional_401k_contributions: 35_750
+    traditional_401k_contributions_desired: 35_750
   output:
     elective_deferral_limit: 32_500
-    capped_traditional_401k_contributions: 32_500
+    traditional_401k_contributions: 32_500
