+++ policyengine_us/tests/policy/baseline/household/expense/retirement/ira_contribution_limit.yaml
@@ -4,25 +4,25 @@
     people:
       person1:
         age: 40
-        traditional_ira_contributions: 4_000
-        roth_ira_contributions: 2_000
+        traditional_ira_contributions_desired: 4_000
+        roth_ira_contributions_desired: 2_000
   output:
     ira_contribution_limit: 7_000
-    capped_traditional_ira_contributions: 4_000
-    capped_roth_ira_contributions: 2_000
+    traditional_ira_contributions: 4_000
+    roth_ira_contributions: 2_000
 
-- name: Case 2, IRA contributions above the base limit are capped proportionally.
+- name: Case 2, IRA contributions above the base limit are allocated proportionally.
   period: 2024
   input:
     people:
       person1:
         age: 40
-        traditional_ira_contributions: 9_000
-        roth_ira_contributions: 3_000
+        traditional_ira_contributions_desired: 9_000
+        roth_ira_contributions_desired: 3_000
   output:
     ira_contribution_limit: 7_000
-    capped_traditional_ira_contributions: 5_250
-    capped_roth_ira_contributions: 1_750
+    traditional_ira_contributions: 5_250
+    roth_ira_contributions: 1_750
 
 - name: Case 3, IRA catch-up contribution raises the combined limit.
   period: 2024
@@ -30,35 +30,33 @@
     people:
       person1:
         age: 50
-        traditional_ira_contributions: 9_000
-        roth_ira_contributions: 3_000
+        traditional_ira_contributions_desired: 9_000
+        roth_ira_contributions_desired: 3_000
   output:
     ira_contribution_limit: 8_000
-    capped_traditional_ira_contributions: 6_000
-    capped_roth_ira_contributions: 2_000
+    traditional_ira_contributions: 6_000
+    roth_ira_contributions: 2_000
 
-- name: Case 4, desired IRA variables override legacy inputs.
+- name: Case 4, desired IRA variables are limited proportionally.
   period: 2024
   input:
     people:
       person1:
         age: 45
-        traditional_ira_contributions: 100
-        roth_ira_contributions: 100
         traditional_ira_contributions_desired: 2_000
         roth_ira_contributions_desired: 8_000
   output:
-    capped_traditional_ira_contributions: 1_400
-    capped_roth_ira_contributions: 5_600
+    traditional_ira_contributions: 1_400
+    roth_ira_contributions: 5_600
 
-- name: Case 5, federal above-the-line deductions use capped traditional IRA contributions.
+- name: Case 5, federal above-the-line deductions use traditional IRA contributions.
   period: 2024
   input:
     people:
       person1:
         age: 40
         employment_income: 100_000
-        traditional_ira_contributions: 12_000
+        traditional_ira_contributions_desired: 12_000
     tax_units:
       tax_unit:
         members: [person1]
@@ -66,26 +64,26 @@
     above_the_line_deductions: 7_000
     adjusted_gross_income: 93_000
 
-- name: Case 6, Saver's Credit qualified contributions use capped IRA contributions.
+- name: Case 6, Saver's Credit qualified contributions use IRA contributions.
   period: 2024
   input:
     people:
       person1:
         age: 40
-        traditional_ira_contributions: 9_000
-        roth_ira_contributions: 3_000
+        traditional_ira_contributions_desired: 9_000
+        roth_ira_contributions_desired: 3_000
   output:
     savers_credit_qualified_contributions: 7_000
 
-- name: Case 7, Puerto Rico retirement deduction uses capped IRA contributions.
+- name: Case 7, Puerto Rico retirement deduction uses IRA contributions.
   period: 2024
   input:
     people:
       person1:
         age: 40
         pr_agi_person: 10_000
-        traditional_ira_contributions: 9_000
-        roth_ira_contributions: 3_000
+        traditional_ira_contributions_desired: 9_000
+        roth_ira_contributions_desired: 3_000
     households:
       household:
         members: [person1]
