+++ policyengine_us/tests/policy/contrib/tax_exempt/tax_exempt_reform.yaml
@@ -69,6 +69,6 @@
     gov.contrib.tax_exempt.overtime.payroll_tax_exempt: false
     gov.contrib.tax_exempt.tip_income.payroll_tax_exempt: false
     employment_income: 100_000
-    traditional_401k_contributions: 10_000
+    traditional_401k_contributions_desired: 10_000
   output:
     payroll_tax_gross_wages: 100_000
