+++ policyengine_us/variables/household/expense/retirement/traditional_403b_contributions.py
@@ -7,6 +7,25 @@ class traditional_403b_contributions(Variable):
     label = "Traditional 403(b) contributions"
     unit = USD
     documentation = (
-        "Legacy input for desired contributions to traditional 403(b) accounts."
+        "Traditional 403(b) contributions after applying the combined "
+        "401(k) and 403(b) elective deferral limit."
     )
     definition_period = YEAR
+    reference = "https://www.law.cornell.edu/uscode/text/26/402#g"
+
+    def formula(person, period, parameters):
+        raw = person("uncapped_traditional_403b_contributions", period)
+        total_desired = add(
+            person,
+            period,
+            [
+                "uncapped_traditional_401k_contributions",
+                "uncapped_roth_401k_contributions",
+                "uncapped_traditional_403b_contributions",
+                "uncapped_roth_403b_contributions",
+            ],
+        )
+        scale = min_(
+            person("elective_deferral_limit", period) / max_(total_desired, 1), 1
+        )
+        return raw * scale
